requirements.txt
alabaster==0.7.12
Babel==2.6.0
certifi==2018.10.15
chardet==3.0.4
docutils==0.14
idna==2.7
imagesize==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
packaging==18.0
Pygments==2.2.0
pyparsing==2.3.0
pytz==2018.7
requests==2.20.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
urllib3==1.24.1
virtualenv==16.1.0

Flask~=1.1.1
pymongo~=3.10.1
tweepy~=3.8.0
mongomock~=3.19.0
pandas~=1.0.3
matplotlib~=3.2.1
python-dateutil~=2.8.1
wordcloud~=1.7.0
nltk~=3.5
scikit-learn~=0.22.2.post1